requirements.txt
opencv
scikit-image
torchvision
numpy
scipy
h5py
torch
pt-trainer
autoencoder